+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support deleting Flux HelmRelease-managed clusters, including their referenced OCIRepository (if not used by another HelmRelease).
+
 ### Changed
 
 - Go: Update dependencies.